+++ csvEditorHtml/browser/CHANGELOG.md
@@ -4,9 +4,15 @@ note that the changelog is also present on the website when you click on the ver
 
 ### [Unreleased]
 
+## 1.2.3
+
+- (same as vs code extension v0.6.3)
+- fixed issue 70: Removing columns doesn't remove header
+	- undo/redo does not work with column headers
+
 ## 1.2.2
 
-- (same as vs code extension v6.0.2)
+- (same as vs code extension v0.6.2)
 - fixed issue 63: column header cells can now be edited
 - fixed issue 66: added readonly mode
 - fixed issue 64: some non-text keys (e.g. volume controls) not longer clear cell values
@@ -14,12 +20,12 @@ note that the changelog is also present on the website when you click on the ver
 
 ## 1.2.1
 
-- (same as vs code extension v6.0.1)
+- (same as vs code extension v0.6.1)
 - added button to resize column to match their content
 
 ## 1.2.0
 
-- (same as vs code extension v6.0.0)
+- (same as vs code extension v0.6.0)
 - fixed issue where reordering/sorting breaks inserting
 - new logo